yanks timeline event tooltip
